+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2019-08-30
+### Add
+- Firefox add-on use webextension-polyfill
+- Right click icon to open chhoe taigi home page
+- Option page
+### Changed
+- Use webpack + vue js refactor
+
 ## [0.1.6] - 2019-04-10
 ### Fixed
 - Use trim() Removes whitespace from selected text both ends of a string
 ### Changed
-- popup height to 600px
+- Popup height to 600px
 ### Added
 - Keyboard shortcuts for activate the extension